code/roadmap-agent/requirements.txt
fastapi
uvicorn
langchain_community
langchain
openai
matplotlib
plotly
python-dotenv